requirements.txt
certifi==2024.2.2
chardet==3.0.4
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
keyboard==0.13.5
packaging==24.0
pillow==10.3.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
pytesseract==0.3.10
rfc3986==1.5.0
screeninfo==0.8.1
sniffio==1.3.1
tk==0.1.0
